requirements.txt
streamlit
tensorflow
numpy
pillow
opencv-python
pandas
scikit-learn
